+++ src/Node.php
@@ -139,6 +139,8 @@ final class Node
     }
 
     /**
+     * @internal
+     *
      * @return Map<string, string>
      */
     public function attributes(): Map